refactor(plugins/json): create package directory
