+++ plugins/json/pkg/json/json_test.go
@@ -1,4 +1,20 @@
-package main
+/*
+Copyright (C) 2022 The Falco Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package json
 
 import (
 	"fmt"
@@ -101,7 +117,7 @@ func TestExtractValue(t *testing.T) {
 		fieldType: sdk.FieldTypeCharBuf,
 		value:     nil,
 	}
-	e := &MyPlugin{}
+	e := &Plugin{}
 
 	// invalid json pointer
 	testRequest.arg = "invalid_pointer"
@@ -183,7 +199,7 @@ func TestExtractObject(t *testing.T) {
 		fieldType: sdk.FieldTypeCharBuf,
 		value:     nil,
 	}
-	e := &MyPlugin{}
+	e := &Plugin{}
 
 	// extract object with json.obj
 	err = e.Extract(testRequest, testEvent)
@@ -229,7 +245,7 @@ func TestExtractRawtime(t *testing.T) {
 		fieldType: sdk.FieldTypeUint64,
 		value:     nil,
 	}
-	e := &MyPlugin{}
+	e := &Plugin{}
 
 	// extract object with json.rawtime
 	err = e.Extract(testRequest, testEvent)
